% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard_cal.R
\name{Cal.ARI}
\alias{Cal.ARI}
\title{Title Adjust Rank Index}
\usage{
Cal.ARI(df, col1, col2)
}
\arguments{
\item{df}{input data frame}

\item{col1}{name of interest variable 1 column in df}

\item{col2}{name of interest variable 2 column in df}
}
\value{
adjust ARI value
}
\description{
Title Adjust Rank Index
}
