% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.auc.R
\name{cal.auc}
\alias{cal.auc}
\title{Title}
\usage{
cal.auc(consencus.result, k)
}
\arguments{
\item{consencus.result}{the list generated by \code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}}.}

\item{k}{cluster number.}
}
\value{
Area under the curve.
}
\description{
Title
}
