% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid2exp.R
\name{centroid2exp}
\alias{centroid2exp}
\title{Title To predict Samples cluter based on their centroid}
\usage{
centroid2exp(centroid, vd)
}
\arguments{
\item{centroid}{a centroid calculated by crossICC}

\item{vd}{a expression dataframe to predict clusters, with featurer in rows and samples in colomns,(no NA permitted )}
}
\description{
Title To predict Samples cluter based on their centroid
}
\examples{
\donttest{
crossICC.object <- CrossICC(example.matrices, max.iter = 5, use.shiny = FALSE)
train.centroid <- cluster.centroid(crossICC.object[[2]]$platforms[[1]],
crossICC.object[[2]]$gene.signature,crossICC.object[[2]]$clusters$clusters)
data(test.data)
predict.result<-centroid2exp(train.centroid,test.data)
}
}
