% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{centroidOfcentroid}
\alias{centroidOfcentroid}
\title{Title Return centroid of centroid from each plat form}
\usage{
centroidOfcentroid(centroid.list, cluster)
}
\arguments{
\item{centroid.list}{a list stored the centroid}

\item{model}{a CrossICC Object}
}
\value{
a list contains a vecter that store the predict clusters and a normalized expression matrix
}
\description{
Title Return centroid of centroid from each plat form
}
