% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.centroid.R
\name{cluster.centroid}
\alias{cluster.centroid}
\title{Title get centroids of clusters by a set of gene signature and their cluster information}
\usage{
cluster.centroid(x, gene.signature, cluster)
}
\arguments{
\item{x}{expression matrix for calculating centroid}

\item{gene.signature}{a vector of genenames representing a certain siganture of pathway or biological function.}

\item{cluster}{a named cluster numbers with gene name as the cluster name and the cluster number as the value}
}
\description{
Title get centroids of clusters by a set of gene signature and their cluster information
}
