% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.feature.R
\name{com.feature}
\alias{com.feature}
\title{Title Extract Common Features}
\usage{
com.feature(x, y, z, method = "merge")
}
\arguments{
\item{x}{a vector, i.e., should be x[[1]] if x is a column of data.table/data.frame.}

\item{y}{ditto.}

\item{z}{ditto.}

\item{method}{should be "merge" or "overlap".}
}
\description{
Title Extract Common Features
}
