% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossICC.R
\docType{package}
\name{CrossICC}
\alias{CrossICC}
\alias{CrossICC-package}
\title{CrossICC: Automatically Aggregating and Summarizing Bioinformatics Results for Interactive Report.}
\description{
The CrossICC package provides three categories of important functions:
foo, bar and baz.
}
\section{CrossICC functions}{

The foo functions ...
}

