% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossICC.R
\name{CrossICCInput}
\alias{CrossICCInput}
\title{Read file into CrossICC input}
\usage{
CrossICCInput(files)
}
\arguments{
\item{files}{a list for filenames, usually a returned value of list.files() function}
}
\value{
list contains matrices from each platform parsing from file provided .
}
\description{
Read file into CrossICC input
}
\examples{
files <- list.files(path=".", pattern = '.csv')
CrossICC.input <- CrossICCInput(files)
}
