% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo.platforms}
\alias{demo.platforms}
\title{list containing several different (eSet) matrix.}
\format{\emph{Note: this is format of single matrix.} A matrix with rows of features (genes) and column of samples:
\describe{
\item{sample1}{GSM*}
\item{sample2}{GSM*}
...
}}
\usage{
demo.platforms
}
\description{
each matrix comes from GSE file with feature names.
}
\keyword{datasets}
