% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.matrices}
\alias{example.matrices}
\title{list containing 3 different (eSet) matrix.}
\format{\emph{Note: this is format of single matrix.} A matrix with rows of features (genes) and column of samples:
\describe{
\item{sample1}{GSM*}
\item{sample2}{GSM*}
...
}}
\usage{
example.matrices
}
\description{
each matrix comes from GSE file with artifical feature names.
}
\keyword{datasets}
