% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard_cal.R
\name{get_jarrad_index_df_fromDF}
\alias{get_jarrad_index_df_fromDF}
\title{Title get jaccard index of list factors}
\usage{
get_jarrad_index_df_fromDF(df1, df2, universe = NULL)
}
\arguments{
\item{df1}{an annotated data frame with cluster at the seccond column}

\item{df2}{an annotated data frame with cluster at the seccond column}

\item{universe}{(Optional) total number of all strings that vec1 and vec2 comes from}
}
\value{
a data frame of Jaccard index or a list contains two dataframe (jaccard index and Fisher's test P value list )
}
\description{
Title get jaccard index of list factors
}
