% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard_cal.R
\name{get_jarrad_index_df_fromlist}
\alias{get_jarrad_index_df_fromlist}
\title{Title get jaccard index of list factors}
\usage{
get_jarrad_index_df_fromlist(list1, list2, universe = NULL)
}
\arguments{
\item{list1}{a list that contains a lot vectors}

\item{list2}{a list that contains a lot vectors}

\item{universe}{(Optional) total number of all strings that vec1 and vec2 comes from}
}
\value{
a data frame of Jaccard index or a list contains two dataframe (jaccard index and Fisher's test P value list )
}
\description{
Title get jaccard index of list factors
}
