% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard_cal.R
\name{get_overlap_test_by_fisher}
\alias{get_overlap_test_by_fisher}
\title{Title Get fihser test p value for overlaps}
\usage{
get_overlap_test_by_fisher(vec1, vec2, universe)
}
\arguments{
\item{vec1}{a string vector}

\item{vec2}{a string vector}

\item{universe}{total number of all strings that vec1 and vec2 comes from}
}
\value{
a P value
}
\description{
Title Get fihser test p value for overlaps
}
