% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshiny.R
\name{runShinyCrossICC}
\alias{runShinyCrossICC}
\title{Run the default shiny report independently}
\usage{
runShinyCrossICC()
}
\description{
\code{runShinyCrossICC} run shiny independently
}
\author{
Qi Zhao
}
