% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA.R
\name{ssGSEA}
\alias{ssGSEA}
\title{Title To get GSEA-like ranked matrix from CrossICC result.}
\usage{
ssGSEA(x, gene.signature, geneset2gene)
}
\arguments{
\item{x}{a eSet object or eSet-like matrix.}

\item{gene.signature}{gene signatures calculated by CrossICC.}

\item{geneset2gene}{a matrix contains geneset (cluster name) mapping to gene.}
}
\value{
a matrix with samples' eigenvalue in different super clusters.
}
\description{
Title To get GSEA-like ranked matrix from CrossICC result.
}
\examples{
\donttest{
CrossICC.object <- CrossICC(demo.platforms, skip.mfs = TRUE, max.iter = 1)
ssGSEA(demo.platforms[[1]], CrossICC.object$gene.signature, CrossICC.object$unioned.genesets)
}
}
