% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossICC.R
\name{summary.CrossICC}
\alias{summary.CrossICC}
\title{Title Summary the CrossICC-returned list to produce human-readable output}
\usage{
summary.CrossICC(result)
}
\arguments{
\item{result}{list-type CrossICC's return value.}
}
\value{
list contains: a matrix of genesets mapping to genes; a named atomic vetor of samples mapping to super clusters.
}
\description{
Title Summary the CrossICC-returned list to produce human-readable output
}
\examples{
\donttest{
CrossICC.object <- CrossICC(example.matrices, max.iter = 20)
summary.CrossICC(CrossICC.object)
}
}
