% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossICC.R
\name{summaryCrossICC}
\alias{summaryCrossICC}
\title{Title Summary the CrossICC-returned list to produce human-readable output}
\usage{
summaryCrossICC(result)
}
\arguments{
\item{result}{list-type CrossICC's return value.}
}
\value{
list contains: a matrix of genesets mapping to genes; a named atomic vetor of samples mapping to super clusters.
}
\description{
Title Summary the CrossICC-returned list to produce human-readable output
}
\examples{
\donttest{
CrossICC.object <- CrossICC(demo.platforms, max.iter = 2)
summaryCrossICC(CrossICC.object)
}
}
