% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{eSet matrix.}
\format{An object of class \code{matrix} with 17355 rows and 9 columns.}
\usage{
x
}
\description{
A matrix from GSE file with artifical feature names.
}
\keyword{datasets}
