% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y}
\alias{y}
\title{eSet matrix.}
\format{An object of class \code{matrix} with 4487 rows and 24 columns.}
\usage{
y
}
\description{
A matrix from GSE file with artifical feature names.
}
\keyword{datasets}
