% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{z}
\alias{z}
\title{eSet matrix.}
\format{An object of class \code{matrix} with 4264 rows and 11 columns.}
\usage{
z
}
\description{
A matrix from GSE file with artifical feature names.
}
\keyword{datasets}
