% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{centroidOfcentroid}
\alias{centroidOfcentroid}
\title{Title Return centroid of centroid from each platform}
\usage{
centroidOfcentroid(centroid.list, cluster)
}
\arguments{
\item{centroid.list}{a list stored the centroid}

\item{cluster}{a named vector with gene names as name and the cluster number as vector value}
}
\value{
a list contains a vecter that store the predict clusters and a normalized expression matrix
}
\description{
Title Return centroid of centroid from each platform
}
