% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossICC.R
\docType{package}
\name{CrossICC}
\alias{CrossICC}
\alias{CrossICC-package}
\title{CrossICC: Automatically Aggregating and Summarizing Bioinformatics Results for Interactive Report.}
\usage{
CrossICC(..., study.names, filter.cutoff = 0.5, fdr.cutoff = 0.1,
  output.dir = NULL, max.K = 10, max.iter = 20, rep.runs = 1000,
  pItem = 0.8, pFeature = 1, clusterAlg = "hc",
  distance = "euclidean", sil.filter = "soft",
  heatmap.order = "up.based", cc.seed = 5000, cluster.cutoff = 0.05,
  ebayes.cutoff = 0.1, ebayes.mode = "both", cross = "sample",
  skip.merge.dup = FALSE, skip.mm = FALSE, skip.mfs = FALSE,
  use.shiny = TRUE)
}
\arguments{
\item{...}{all datasets (matrices is better)}

\item{study.names}{a vector containing all study names}

\item{filter.cutoff}{cutoff value during mad filtering.}

\item{fdr.cutoff}{cutoff value during fdr filtering.}

\item{output.dir}{the results' output directory.}

\item{max.K}{the maximum cluster number of ConsensusClusterPlus.}

\item{max.iter}{the maximum number of iterations.}

\item{rep.runs}{number of subsamples during clustering.}

\item{pItem}{proportion of items to sample during clustering.}

\item{pFeature}{proportion of features to sample during clustering.}

\item{clusterAlg}{cluster algorithm. Could be 'hc' heirarchical (hclust),
'pam' for paritioning around medoids, 'km' for k-means upon data matrix,
'kmdist' for k-means upon distance matrices (former km option),
or a function that returns a clustering.}

\item{distance}{Could be 'pearson': (1 - Pearson correlation),
'spearman' (1 - Spearman correlation),
'euclidean', 'binary', 'maximum', 'canberra', 'minkowski" or custom distance function.}

\item{sil.filter}{silhouetee width filtering mode. Could be "soft" or "hard". If "hard", all negtive silhouetee width value will be set to 0. Default is "soft" (to do nothing).}

\item{heatmap.order}{gene order for heatmaps. Default is "up.based", with which genes will be arranged as up-regulated order in super-clusters across all matrices. Or can be set to "concordant" for all in same order.}

\item{cc.seed}{sets random seed for reproducible results.}

\item{cluster.cutoff}{cutoff value during determining cluster numbers.}

\item{ebayes.cutoff}{p-value cutoff when select differentially expressed probes.}

\item{ebayes.mode}{'up' or 'both'. Choose only up-regulated genes or all differentially expressed genes.}

\item{cross}{cross analysis object. Could be "cluster" for clusters, "sample" for samples or "none" (used for single dataset).}

\item{skip.merge.dup}{skip merge multiple probes for one gene (duplicates) or not. Default is FALSE (not skip).}

\item{skip.mm}{skip call MergeMaid process or not. Default is FALSE (not skip).}

\item{use.shiny}{if TRUE, a shiny app will appear after running this main function.}
}
\value{
A nested list with iteration time as its name and list containing consensus cluster,
gene signature and balanced cluster as its value.
}
\description{
The CrossICC package provides three categories of important functions:
foo, bar and baz.
}
\section{CrossICC functions}{

The foo functions ...
}

\examples{
\donttest{
# It takes too long time for running code below, so ignore them in R CMD check.
CrossICC.obj <- CrossICC(demo.platforms, skip.mfs = TRUE, max.iter = 100, use.shiny = FALSE, cross = "cluster", fdr.cutoff = 0.1, ebayes.cutoff = 0.1, filter.cutoff = 0.1)
}
}
\seealso{
\code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}}
}
