% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossICC.R
\docType{package}
\name{CrossICC}
\alias{CrossICC}
\alias{CrossICC-package}
\title{CrossICC: Automatically Aggregating and Summarizing Bioinformatics Results for Interactive Report.}
\usage{
CrossICC(..., study.names, filter.cutoff = 0.5, fdr.cutoff = 0.001,
  output.dir = "~/", max.K = 10, max.iter = 20, rep.runs = 1000,
  n.platform = 2, pItem = 0.8, pFeature = 1, clusterAlg = "hc",
  distance = "euclidean", sil.filter = "soft",
  heatmap.order = "up.based", com.mode = "overlap", cc.seed = NULL,
  cluster.cutoff = 0.05, ebayes.cutoff = 0.1, ebayes.mode = "up",
  cross = "cluster", skip.merge.dup = TRUE, skip.mm = FALSE,
  skip.mfs = FALSE, use.shiny = TRUE)
}
\arguments{
\item{...}{all datasets (matrices is better)}

\item{study.names}{a vector containing all study names}

\item{filter.cutoff}{low variability (median absolute deviation (MAD)) cutoff threshold, default is 0.5.}

\item{fdr.cutoff}{cutoff value during fdr filtering.}

\item{output.dir}{the results' output directory.}

\item{max.K}{the maximum cluster number of ConsensusClusterPlus.}

\item{max.iter}{the maximum number of iterations.}

\item{rep.runs}{number of subsamples during clustering.}

\item{n.platform}{to filter the signature with it's meta-cluster group in platforms.
That is, if the parameter is set to 2 (default),
the signature (like hgnc symbol ESR1) in a certain meta-cluser (like K1) must exists more than 2 times among data of all platforms;
otherwise, it will not be reported.}

\item{pItem}{proportion of items to sample during clustering.}

\item{pFeature}{proportion of features to sample during clustering.}

\item{clusterAlg}{cluster algorithm. Could be 'hc' heirarchical (hclust),
'pam' for paritioning around medoids, 'km' for k-means upon data matrix,
'kmdist' for k-means upon distance matrices (former km option),
or a function that returns a clustering.}

\item{distance}{Could be 'pearson': (1 - Pearson correlation),
'spearman' (1 - Spearman correlation),
'euclidean', 'canberra', 'minkowski" or custom distance function.}

\item{sil.filter}{silhouetee width filtering mode. Could be "soft" or "hard". If "hard", all negtive silhouetee width value will be set to 0. Default is "soft" (to do nothing).}

\item{heatmap.order}{gene order for heatmaps. Default is "up.based", with which genes will be arranged as up-regulated order in meta-clusters across all matrices. Or can be set to "concordant" for all in same order.}

\item{com.mode}{mode for choose common features when pre-processing data. Could be "overlap" (use intersection, default) or "merge" (keep all features).}

\item{cc.seed}{sets random seed for reproducible results.}

\item{cluster.cutoff}{cutoff value during determining cluster numbers.}

\item{ebayes.cutoff}{p-value cutoff when select differentially expressed probes.}

\item{ebayes.mode}{'up' or 'both'. Choose only up-regulated genes or all differentially expressed genes when determining MDEGs. default is 'up'}

\item{cross}{object type when determining meta-cluster. Could be "cluster" for clusters by ConsencusClusterPlus, "sample" for samples or "none" (only used for single dataset).}

\item{skip.merge.dup}{skip merge multiple probes for one gene (duplicates) or not. Default is TRUE (it is highly recommended that user has their data pre-processed well).}

\item{skip.mm}{skip MergeMaid processing or not. Default is FALSE (not skip).}

\item{skip.mfs}{by default, the datasets will be normalized at the start,
and the genes or features that have no or few contributions to the final clusters will be filtered out. To skip this process, you can set this parameter to TRUE. Only try when you're sure that you're working with pre-processed datasets.}

\item{use.shiny}{if TRUE, a shiny app will appear after running this main function.}
}
\value{
A nested list with iteration time as its name and list containing consensus cluster,
gene signature and balanced cluster as its value.
}
\description{
The CrossICC package provides three categories of important functions:
foo, bar and baz.
}
\section{CrossICC functions}{

The foo functions ...
}

\examples{
\donttest{
# It takes too long time for running code below, so ignore them in R CMD check.
CrossICC.obj <- CrossICC(demo.platforms, skip.mfs = TRUE, max.iter = 2)
}
}
\seealso{
\code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}}
}
