% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predictor}
\alias{predictor}
\title{To calculate the correlation between the predictor centroid and the validation centroid.}
\usage{
predictor(pre.dat, model)
}
\arguments{
\item{pre.dat}{a eSet object or eSet-like matrix with features in rows and samples in columns}

\item{model}{a list containing CrossICC result}
}
\value{
a list contains a vecter that store the predict clusters and a normalized expression matrix
}
\description{
To calculate the correlation between the predictor centroid and the validation centroid.
}
\examples{
data("demo.platforms")
CrossICC.object <- CrossICC(demo.platforms, skip.mfs = TRUE, max.iter = 1, overwrite = TRUE)
predicted <- predictor(demo.platforms[[1]], CrossICC.object)
}
