% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA.R
\name{ssGSEA}
\alias{ssGSEA}
\title{Title}
\usage{
ssGSEA(x, gene.signature, geneset2gene, plot.file = FALSE,
  color = c("purple", "#910202", "#035005", "#1273d6", "#ff8207"), cluster)
}
\arguments{
\item{color}{}
}
\description{
Title
}
