% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA.R
\name{ssGSEA}
\alias{ssGSEA}
\title{To get GSEA-like ranked matrix from CrossICC result.}
\usage{
ssGSEA(x, gene.signature, geneset2gene, cluster)
}
\arguments{
\item{x}{a eSet object or eSet-like matrix.}

\item{gene.signature}{gene signatures calculated by CrossICC.}

\item{geneset2gene}{a matrix contains geneset (cluster name) mapping to gene.}

\item{cluster}{CrossICC returned clusters. Note: Must mapping to x!}
}
\value{
a matrix with samples' eigenvalue in different super clusters.
}
\description{
To get GSEA-like ranked matrix from CrossICC result.
}
\examples{
data("demo.platforms")
CrossICC.object <- CrossICC(demo.platforms, skip.mfs = TRUE, max.iter = 1, overwrite = TRUE, outdir = tempdir())
Mcluster <- paste("K", CrossICC.object$clusters$clusters[[1]], sep = "")
CrossICC.ssGSEA <- ssGSEA(x = demo.platforms[[1]], gene.signature = CrossICC.object$gene.signature,
geneset2gene = CrossICC.object$unioned.genesets, cluster = Mcluster)
}
