% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA.R
\name{ssGSEA}
\alias{ssGSEA}
\title{Title To get GSEA-like ranked matrix from CrossICC result.}
\usage{
ssGSEA(x, gene.signature, geneset2gene, color = c("purple", "#910202",
  "#035005", "#1273d6", "#ff8207"), cluster)
}
\arguments{
\item{x}{a eSet object or eSet-like matrix.}

\item{gene.signature}{gene signatures calculated by CrossICC.}

\item{geneset2gene}{a matrix contains geneset (cluster name) mapping to gene.}

\item{color}{to be deprecated.}

\item{cluster}{only used when plot.file is provided. Single super cluster vector of list from CrossICC result list.}
}
\value{
a matrix with samples' eigenvalue in different super clusters.
}
\description{
Title To get GSEA-like ranked matrix from CrossICC result.
}
\examples{
\donttest{
CrossICC.object <- CrossICC(demo.platforms, max.iter = 2)
ssGSEA(demo.platforms[[1]], CrossICC.object$gene.signature, CrossICC.object$unioned.genesets)
}
}
