% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossICC.R
\name{summaryCrossICC}
\alias{summaryCrossICC}
\title{Summary the CrossICC-returned list to produce human-readable output}
\usage{
summaryCrossICC(result)
}
\arguments{
\item{result}{list-type CrossICC's return value.}
}
\value{
list contains: a matrix of genesets mapping to genes; a named atomic vetor of samples mapping to super clusters.
}
\description{
Summary the CrossICC-returned list to produce human-readable output
}
\examples{
data("demo.platforms")
CrossICC.object <- CrossICC(demo.platforms, skip.mfs = TRUE, max.iter = 1, overwrite = TRUE, output.dir = tempdir())
CrossICC.summary <- summaryCrossICC(CrossICC.object)
}
